# RealTimeChessAlphaSeven.swagger.json
{"swagger":"2.0","info":{"version":"V1","title":"RealTimeChess API"},"paths":{"/Account/Login":{"get":{"tags":["Account"],"operationId":"AccountLoginGet","consumes":[],"produces":[],"parameters":[{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountLoginPost","consumes":[],"produces":[],"parameters":[{"name":"Email","in":"query","required":true,"type":"string"},{"name":"Password","in":"query","required":true,"type":"string"},{"name":"RememberMe","in":"query","required":true,"type":"boolean"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/LoginWith2fa":{"get":{"tags":["Account"],"operationId":"AccountLoginWith2faGet","consumes":[],"produces":[],"parameters":[{"name":"rememberMe","in":"query","required":true,"type":"boolean"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountLoginWith2faPost","consumes":[],"produces":[],"parameters":[{"name":"TwoFactorCode","in":"query","required":true,"type":"string"},{"name":"RememberMachine","in":"query","required":true,"type":"boolean"},{"name":"RememberMe","in":"query","required":true,"type":"boolean"},{"name":"rememberMe","in":"query","required":true,"type":"boolean"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/LoginWithRecoveryCode":{"get":{"tags":["Account"],"operationId":"AccountLoginWithRecoveryCodeGet","consumes":[],"produces":[],"parameters":[{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountLoginWithRecoveryCodePost","consumes":[],"produces":[],"parameters":[{"name":"RecoveryCode","in":"query","required":true,"type":"string"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/Lockout":{"get":{"tags":["Account"],"operationId":"AccountLockoutGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Account/Register":{"get":{"tags":["Account"],"operationId":"AccountRegisterGet","consumes":[],"produces":[],"parameters":[{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountRegisterPost","consumes":[],"produces":[],"parameters":[{"name":"Email","in":"query","required":true,"type":"string"},{"name":"Password","in":"query","required":true,"type":"string"},{"name":"ConfirmPassword","in":"query","required":false,"type":"string"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/Logout":{"post":{"tags":["Account"],"operationId":"AccountLogoutPost","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Account/ExternalLogin":{"post":{"tags":["Account"],"operationId":"AccountExternalLoginPost","consumes":[],"produces":[],"parameters":[{"name":"provider","in":"query","required":false,"type":"string"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ExternalLoginCallback":{"get":{"tags":["Account"],"operationId":"AccountExternalLoginCallbackGet","consumes":[],"produces":[],"parameters":[{"name":"returnUrl","in":"query","required":false,"type":"string"},{"name":"remoteError","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ExternalLoginConfirmation":{"post":{"tags":["Account"],"operationId":"AccountExternalLoginConfirmationPost","consumes":[],"produces":[],"parameters":[{"name":"Email","in":"query","required":true,"type":"string"},{"name":"returnUrl","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ConfirmEmail":{"get":{"tags":["Account"],"operationId":"AccountConfirmEmailGet","consumes":[],"produces":[],"parameters":[{"name":"userId","in":"query","required":false,"type":"string"},{"name":"code","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ForgotPassword":{"get":{"tags":["Account"],"operationId":"AccountForgotPasswordGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountForgotPasswordPost","consumes":[],"produces":[],"parameters":[{"name":"Email","in":"query","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ForgotPasswordConfirmation":{"get":{"tags":["Account"],"operationId":"AccountForgotPasswordConfirmationGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Account/ResetPassword":{"get":{"tags":["Account"],"operationId":"AccountResetPasswordGet","consumes":[],"produces":[],"parameters":[{"name":"code","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Account"],"operationId":"AccountResetPasswordPost","consumes":[],"produces":[],"parameters":[{"name":"Email","in":"query","required":true,"type":"string"},{"name":"Password","in":"query","required":true,"type":"string"},{"name":"ConfirmPassword","in":"query","required":false,"type":"string"},{"name":"Code","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Account/ResetPasswordConfirmation":{"get":{"tags":["Account"],"operationId":"AccountResetPasswordConfirmationGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Account/AccessDenied":{"get":{"tags":["Account"],"operationId":"AccountAccessDeniedGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/api/ChessMatches":{"get":{"tags":["ChessMatches"],"operationId":"ApiChessMatchesGet","consumes":[],"produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/ChessMatch"}}}}},"post":{"tags":["ChessMatches"],"operationId":"ApiChessMatchesPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"chessMatch","in":"body","required":false,"schema":{"$ref":"#/definitions/ChessMatch"}}],"responses":{"200":{"description":"Success"}}}},"/api/ChessMatches/{id}":{"get":{"tags":["ChessMatches"],"operationId":"ApiChessMatchesByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["ChessMatches"],"operationId":"ApiChessMatchesByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"chessMatch","in":"body","required":false,"schema":{"$ref":"#/definitions/ChessMatch"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["ChessMatches"],"operationId":"ApiChessMatchesByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/Manage/Index":{"get":{"tags":["Manage"],"operationId":"ManageIndexGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Manage"],"operationId":"ManageIndexPost","consumes":[],"produces":[],"parameters":[{"name":"Username","in":"query","required":false,"type":"string"},{"name":"IsEmailConfirmed","in":"query","required":true,"type":"boolean"},{"name":"Email","in":"query","required":true,"type":"string"},{"name":"PhoneNumber","in":"query","required":false,"type":"string"},{"name":"StatusMessage","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/SendVerificationEmail":{"post":{"tags":["Manage"],"operationId":"ManageSendVerificationEmailPost","consumes":[],"produces":[],"parameters":[{"name":"Username","in":"query","required":false,"type":"string"},{"name":"IsEmailConfirmed","in":"query","required":true,"type":"boolean"},{"name":"Email","in":"query","required":true,"type":"string"},{"name":"PhoneNumber","in":"query","required":false,"type":"string"},{"name":"StatusMessage","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/ChangePassword":{"get":{"tags":["Manage"],"operationId":"ManageChangePasswordGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Manage"],"operationId":"ManageChangePasswordPost","consumes":[],"produces":[],"parameters":[{"name":"OldPassword","in":"query","required":true,"type":"string"},{"name":"NewPassword","in":"query","required":true,"type":"string"},{"name":"ConfirmPassword","in":"query","required":false,"type":"string"},{"name":"StatusMessage","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/SetPassword":{"get":{"tags":["Manage"],"operationId":"ManageSetPasswordGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Manage"],"operationId":"ManageSetPasswordPost","consumes":[],"produces":[],"parameters":[{"name":"NewPassword","in":"query","required":true,"type":"string"},{"name":"ConfirmPassword","in":"query","required":false,"type":"string"},{"name":"StatusMessage","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/ExternalLogins":{"get":{"tags":["Manage"],"operationId":"ManageExternalLoginsGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/LinkLogin":{"post":{"tags":["Manage"],"operationId":"ManageLinkLoginPost","consumes":[],"produces":[],"parameters":[{"name":"provider","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/LinkLoginCallback":{"get":{"tags":["Manage"],"operationId":"ManageLinkLoginCallbackGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/RemoveLogin":{"post":{"tags":["Manage"],"operationId":"ManageRemoveLoginPost","consumes":[],"produces":[],"parameters":[{"name":"LoginProvider","in":"query","required":false,"type":"string"},{"name":"ProviderKey","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/TwoFactorAuthentication":{"get":{"tags":["Manage"],"operationId":"ManageTwoFactorAuthenticationGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/Disable2faWarning":{"get":{"tags":["Manage"],"operationId":"ManageDisable2faWarningGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/Disable2fa":{"post":{"tags":["Manage"],"operationId":"ManageDisable2faPost","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/EnableAuthenticator":{"get":{"tags":["Manage"],"operationId":"ManageEnableAuthenticatorGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Manage"],"operationId":"ManageEnableAuthenticatorPost","consumes":[],"produces":[],"parameters":[{"name":"Code","in":"query","required":true,"type":"string"},{"name":"SharedKey","in":"query","required":false,"type":"string"},{"name":"AuthenticatorUri","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/Manage/ShowRecoveryCodes":{"get":{"tags":["Manage"],"operationId":"ManageShowRecoveryCodesGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/ResetAuthenticatorWarning":{"get":{"tags":["Manage"],"operationId":"ManageResetAuthenticatorWarningGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/ResetAuthenticator":{"post":{"tags":["Manage"],"operationId":"ManageResetAuthenticatorPost","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/GenerateRecoveryCodesWarning":{"get":{"tags":["Manage"],"operationId":"ManageGenerateRecoveryCodesWarningGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/Manage/GenerateRecoveryCodes":{"post":{"tags":["Manage"],"operationId":"ManageGenerateRecoveryCodesPost","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}},"/api/MatchPlayers":{"get":{"tags":["MatchPlayers"],"operationId":"ApiMatchPlayersGet","consumes":[],"produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/MatchPlayer"}}}}},"post":{"tags":["MatchPlayers"],"operationId":"ApiMatchPlayersPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"matchPlayer","in":"body","required":false,"schema":{"$ref":"#/definitions/MatchPlayer"}}],"responses":{"200":{"description":"Success"}}}},"/api/MatchPlayers/{id}":{"get":{"tags":["MatchPlayers"],"operationId":"ApiMatchPlayersByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["MatchPlayers"],"operationId":"ApiMatchPlayersByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"matchPlayer","in":"body","required":false,"schema":{"$ref":"#/definitions/MatchPlayer"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["MatchPlayers"],"operationId":"ApiMatchPlayersByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Moves":{"get":{"tags":["Moves"],"operationId":"ApiMovesGet","consumes":[],"produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/Move"}}}}},"post":{"tags":["Moves"],"operationId":"ApiMovesPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"move","in":"body","required":false,"schema":{"$ref":"#/definitions/Move"}}],"responses":{"200":{"description":"Success"}}}},"/api/Moves/{id}":{"get":{"tags":["Moves"],"operationId":"ApiMovesByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["Moves"],"operationId":"ApiMovesByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"move","in":"body","required":false,"schema":{"$ref":"#/definitions/Move"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Moves"],"operationId":"ApiMovesByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Players":{"get":{"tags":["Players"],"operationId":"ApiPlayersGet","consumes":[],"produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/Player"}}}}},"post":{"tags":["Players"],"operationId":"ApiPlayersPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["application/json"],"parameters":[{"name":"player","in":"body","required":false,"schema":{"$ref":"#/definitions/Player"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/IActionResult"}}}}},"/api/Players/{id}":{"get":{"tags":["Players"],"operationId":"ApiPlayersByIdGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Player"}}}},"put":{"tags":["Players"],"operationId":"ApiPlayersByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"player","in":"body","required":false,"schema":{"$ref":"#/definitions/Player"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Players"],"operationId":"ApiPlayersByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/PlayerTypes":{"get":{"tags":["PlayerTypes"],"operationId":"ApiPlayerTypesGet","consumes":[],"produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/PlayerType"}}}}},"post":{"tags":["PlayerTypes"],"operationId":"ApiPlayerTypesPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"playerType","in":"body","required":false,"schema":{"$ref":"#/definitions/PlayerType"}}],"responses":{"200":{"description":"Success"}}}},"/api/PlayerTypes/{id}":{"get":{"tags":["PlayerTypes"],"operationId":"ApiPlayerTypesByIdGet","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["PlayerTypes"],"operationId":"ApiPlayerTypesByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"playerType","in":"body","required":false,"schema":{"$ref":"#/definitions/PlayerType"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["PlayerTypes"],"operationId":"ApiPlayerTypesByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"ChessMatch":{"type":"object","properties":{"chessMatchId":{"format":"int32","type":"integer"},"playerTypeName":{"type":"string"},"matchStartTime":{"format":"date-time","type":"string"},"matchEndTime":{"format":"date-time","type":"string"},"isActive":{"type":"boolean"},"isDeleted":{"type":"boolean"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"deleted":{"format":"date-time","type":"string"},"matchPlayers":{"type":"array","items":{"$ref":"#/definitions/MatchPlayer"}}}},"MatchPlayer":{"type":"object","properties":{"matchPlayerId":{"format":"int32","type":"integer"},"chessMatchId":{"format":"int32","type":"integer"},"playerId":{"format":"int32","type":"integer"},"playerTypeId":{"format":"int32","type":"integer"},"isDeleted":{"type":"boolean"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"deleted":{"format":"date-time","type":"string"},"moves":{"type":"array","items":{"$ref":"#/definitions/Move"}}}},"Move":{"type":"object","properties":{"moveId":{"format":"int32","type":"integer"},"playerTypeName":{"type":"string"},"matchPlayerId":{"format":"int32","type":"integer"},"algebraicChessNotation":{"type":"string"},"gameClockBeginMove":{"format":"date-time","type":"string"},"gameClockEndMove":{"format":"date-time","type":"string"},"positionBeginMove":{"type":"string"},"positionEndMove":{"type":"string"},"isDeleted":{"type":"boolean"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"deleted":{"format":"date-time","type":"string"}}},"Player":{"type":"object","properties":{"playerId":{"format":"int32","type":"integer"},"authenticationId":{"format":"int32","type":"integer"},"firstName":{"type":"string"},"lastName":{"type":"string"},"numWins":{"format":"int32","type":"integer"},"numLosses":{"format":"int32","type":"integer"},"isActive":{"type":"boolean"},"isDeleted":{"type":"boolean"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"},"deleted":{"format":"date-time","type":"string"},"matchPlayers":{"type":"array","items":{"$ref":"#/definitions/MatchPlayer"}}}},"IActionResult":{"type":"object","properties":{}},"PlayerType":{"type":"object","properties":{"playerTypeId":{"format":"int32","type":"integer"},"playerTypeName":{"type":"string"},"turnOrder":{"format":"int32","type":"integer"},"matchPlayers":{"type":"array","items":{"$ref":"#/definitions/MatchPlayer"}}}}},"securityDefinitions":{}}